Add files via upload
Creating Linear Regression without using the library Scikit-learn and Scipy. We create the desired inputs and targets and create a model that produces an output that matches the outputs with the targets. The result shows 2,3 for weights and 5 for biases which is correct. The learning rate(eta) is 0.02. The objective function is L2-norm Loss, an optimal choice for Linear Regression, and the optimization algorithm is the Gradient Descent Update Rules: The variables that we are trying until we find the minimum are the weights and the biases.